#1 - refactoring: extracted WFS20ResourceTestBase
