+++ src/test/java/de/bayern/gdi/WfsTest.java
@@ -20,64 +20,29 @@ package de.bayern.gdi;
 import de.bayern.gdi.services.WFSMeta;
 import de.bayern.gdi.services.WFSMeta.Feature;
 import de.bayern.gdi.services.WFSMetaExtractor;
-import de.bayern.gdi.utils.Config;
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.charset.Charset;
-import junit.framework.TestCase;
-import static net.jadler.Jadler.closeJadler;
-import static net.jadler.Jadler.initJadler;
-
-import static net.jadler.Jadler.onRequest;
-import static net.jadler.Jadler.port;
-import org.apache.commons.io.IOUtils;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
+
+import static net.jadler.Jadler.port;
+import static org.junit.Assert.assertFalse;
+
 /**
- *
  * @author Juergen Weichand
  */
-public class WfsTest extends TestCase {
-
-    /**
-     * Constant for HTTP_OKAY.
-     */
-    public static final int HTTP_OKAY = 200;
+public class WfsTest extends WFS20ResourceTestBase {
 
     /**
      * Constant for FEATURES_PER_PAGE.
      */
     public static final int FEATURES_PER_PAGE = 1037;
 
-    public WfsTest(String testName) {
-        super(testName);
-    }
-
-    @Before
-    @Override
-    public void setUp() throws IOException {
-        try {
-            initJadler();
-        } catch (IllegalStateException ise) {
-            System.out.println("Jadler is already initialized");
-        }
-        Config.initialize(null);
-    }
 
     /**
-     *  The tearDown phase just closes all resources.
-     */
-    @After
-    @Override
-    public void tearDown() {
-        closeJadler();
-    }
-
-    /**
-     *  Test virtuell GeoServer.
-     * @throws IOException Something went wrong
+     * Test virtuell GeoServer.
+     *
+     * @throws IOException        Something went wrong
      * @throws URISyntaxException if URL is wrong
      */
     @Test
@@ -87,8 +52,9 @@ public class WfsTest extends TestCase {
     }
 
     /**
-     *  Test virtuell XtraServer.
-     * @throws IOException Something went wrong
+     * Test virtuell XtraServer.
+     *
+     * @throws IOException        Something went wrong
      * @throws URISyntaxException if URL is wrong
      */
     @Test
@@ -97,24 +63,15 @@ public class WfsTest extends TestCase {
         run("/xtraserver/wfs", "/wfs20/xtraserver/xtraserver-capabilities.xml");
     }
 
-
-
     private void run(String queryPath, String queryResource)
-                throws IOException, URISyntaxException {
+        throws IOException, URISyntaxException {
+        int port = port();
 
-        String body = IOUtils.toString(
-                WfsTest.class.getResourceAsStream(queryResource), "UTF-8"
-        );
-
-        body = body.replace("{DESCRIBESTOREDQUERIES_URL}",
-                buildDescribeStoredQueriesUrl(port()));
-
-        prepareResource(queryPath, body); // prepare GetCapabilities-Response
+        prepareCapabilities(queryResource, queryPath, port);
         prepareDescribeStoredQueries();
 
-
         WFSMeta wfsMeta = new WFSMetaExtractor(
-                buildGetCapabilitiesUrl(queryPath, port())).parse();
+            buildGetCapabilitiesUrl(queryPath, port)).parse();
 
         checkFeatureTypes(wfsMeta);
         checkFeaturesPerPage(wfsMeta);
@@ -154,63 +111,6 @@ public class WfsTest extends TestCase {
         if (fpp == null || !fpp.equals(FEATURES_PER_PAGE)) {
             assertFalse(sb.toString(), true);
         }
-    }
-
-
-    /*
-        prepare Resources
-    */
-    private void prepareResource(String queryPath, String body)
-                throws IOException {
-
-        onRequest()
-                .havingMethodEqualTo("GET")
-                .havingPathEqualTo(queryPath)
-                // .havingBody(isEmptyOrNullString())
-                // .havingHeaderEqualTo("Accept", "application/xml")
-                .respond()
-                 // .withDelay(1, SECONDS)
-                .withStatus(HTTP_OKAY)
-                .withBody(body)
-                .withEncoding(Charset.forName("UTF-8"))
-                .withContentType("application/xml; charset=UTF-8");
-    }
-
-
-    private void prepareDescribeStoredQueries() throws IOException {
-
-        String queryResource = "/wfs20/desc-storedqueries.xml";
-
-        String body = IOUtils.toString(
-                WfsTest.class.getResourceAsStream(queryResource), "UTF-8"
-        );
-
-        prepareResource("/wfs/wfs", body);
-    }
-
-
-    /*
-        URL-Helper
-    */
-    private String buildGetCapabilitiesUrl(String queryPath, int port) {
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("http://localhost:");
-        sb.append(port);
-        sb.append(queryPath);
-        System.out.println("GetCapabilities-URL: " + sb.toString());
-        return sb.toString();
-    }
-
-
-    private String buildDescribeStoredQueriesUrl(int port) {
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("http://localhost:");
-        sb.append(port);
-        sb.append("/wfs/wfs");
-        System.out.println("DescribeStoredQueries-URL: " + sb.toString());
-        return sb.toString();
     }
 
 }
